Redirect to create a producer when there are none
In this commit, we are redirecting the user to create a new producer
whenever they access the application and they have no producer.

For that, we created the producer model that belongs to the user model
and added a new page to host the producer form with a route to it.